Move SmsException to commons module
