--- sms-core/src/main/java/com/github/jackieonway/sms/core/exception/SmsException.java
@@ -1,27 +1,0 @@
-package com.github.jackieonway.sms.core.exception;
-
-/**
- * @author Jackie
- * @version \$id: SmsException.java v 0.1 2019-06-16 18:39 Jackie Exp $$
- */
-public class SmsException extends RuntimeException {
-    private static final String SYS_TYPE_CONFIG_ERROR_MSG = "短信服务商信息配置错误";
-    private static final long serialVersionUID = 3007883817267430838L;
-
-    public SmsException() {
-        super(SYS_TYPE_CONFIG_ERROR_MSG);
-    }
-
-    public SmsException(String message) {
-        super(message);
-    }
-
-    public SmsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public SmsException(Throwable cause) {
-        super(SYS_TYPE_CONFIG_ERROR_MSG, cause);
-    }
-
-}
