+++ sms-core/src/main/java/com/github/jackieonway/sms/core/service/AliSmsService.java
@@ -6,11 +6,11 @@ import com.aliyuncs.IAcsClient;
 import com.aliyuncs.exceptions.ClientException;
 import com.aliyuncs.http.MethodType;
 import com.aliyuncs.profile.DefaultProfile;
+import com.github.jackieonway.sms.commons.exception.SmsException;
 import com.github.jackieonway.sms.commons.utils.GsonUtils;
 import com.github.jackieonway.sms.core.entity.AliSmsRequest;
 import com.github.jackieonway.sms.core.entity.BaseRequest;
 import com.github.jackieonway.sms.core.entity.SmsProperties;
-import com.github.jackieonway.sms.core.exception.SmsException;
 import com.github.jackieonway.sms.core.utls.SmsCacheUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.lang.NonNull;
